+++ lastfm/mockclient/artist_get_info_response.xml
