+++ lastfm/mockclient/artist_get_similar_response.xml
