+++ lastfm/mockclient/get_session_response.xml
