+++ lastfm/mockclient/mockclient.go
