+++ lastfm/mockclient/track_get_similar_response.xml
