+++ lastfm/model.go
