+++ server/ctrladmin/handlers.go
@@ -19,8 +19,8 @@ import (
 	"github.com/nfnt/resize"
 
 	"go.senan.xyz/gonic/db"
+	"go.senan.xyz/gonic/listenbrainz"
 	"go.senan.xyz/gonic/scanner"
-	"go.senan.xyz/gonic/scrobble/listenbrainz"
 	"go.senan.xyz/gonic/transcode"
 )
 
@@ -97,15 +97,7 @@ func (c *Controller) ServeLinkLastFMDo(r *http.Request) *Response {
 	if token == "" {
 		return &Response{code: 400, err: "please provide a token"}
 	}
-	apiKey, err := c.DB.GetSetting(db.LastFMAPIKey)
-	if err != nil {
-		return &Response{redirect: r.Referer(), flashW: []string{fmt.Sprintf("couldn't get api key: %v", err)}}
-	}
-	secret, err := c.DB.GetSetting(db.LastFMSecret)
-	if err != nil {
-		return &Response{redirect: r.Referer(), flashW: []string{fmt.Sprintf("couldn't get secret: %v", err)}}
-	}
-	sessionKey, err := c.lastfmClient.GetSession(apiKey, secret, token)
+	sessionKey, err := c.lastfmClient.GetSession(token)
 	if err != nil {
 		return &Response{
 			redirect: "/admin/home",
